+++ APP/app.py
@@ -8,43 +8,12 @@ app.config['MONGO_DBNAME'] = 'Login'
 app.config['MONGO_URI'] = 'mongodb+srv://ssegroup2:MghlwoCbDqaav9Yc@ssegroup2-vksod.mongodb.net/users?retryWrites=true&w=majority'
 
 mongo = PyMongo(app)
-@app.route('/')
-def index():
-    if 'username' in session:
-        return 'You are logged in as ' + session['username']
 
-    return render_template('index.html')
+app.secret_key = 'mysecret'
 
-@app.route('/login', methods=['POST'])
-
-def login():
-    users = mongo.db.users
-    login_user = users.find_one({'name' : request.form['username']})
-
-    if login_user:
-        if (request.form['pass'], login_user['password'] == login_user['password']):
-            session['username'] = request.form['username']
-            return redirect(url_for('index'))
-
-    return 'Invalid username/password combination'
-
-@app.route('/register', methods=['POST', 'GET'])
-
-def register():
-    if request.method == 'POST':
-        users = mongo.db.users
-        existing_user = users.find_one({'name' : request.form['username']})
-
-        if existing_user is None:
-            hashpass = bcrypt.hashpw(request.form['pass'].encode('utf-8'), bcrypt.gensalt())
-            users.insert({'name' : request.form['username'], 'password' : hashpass})
-            session['username'] = request.form['username']
-            return redirect(url_for('index'))
-        
-        return 'That username already exists!'
-
-    return render_template('register.html')
-
-if __name__ == '__main__':
-    app.secret_key = 'mysecret'
+if __name__=='__main__':
+    # app.secret_key = 'mysecret'
+    from controller_login import *
+    from controller_index import *
+    from controller_register import *
     app.run(debug=True)
