+++ APP/controller_register.py
@@ -1,0 +1,22 @@
+# imports
+from app import app, mongo
+from flask import Flask, render_template, url_for, request, session, redirect
+from flask_pymongo import PyMongo
+import bcrypt
+
+@app.route('/register', methods=['POST', 'GET'])
+
+def register():
+    if request.method == 'POST':
+        users = mongo.db.users
+        existing_user = users.find_one({'name' : request.form['username']})
+
+        if existing_user is None:
+            hashpass = bcrypt.hashpw(request.form['pass'].encode('utf-8'), bcrypt.gensalt())
+            users.insert({'name' : request.form['username'], 'password' : hashpass})
+            session['username'] = request.form['username']
+            return redirect(url_for('register'))
+        
+        return 'That username already exists!'
+
+    return render_template('register.html')
